Refactored printer classes into subdirectories and adjusted include paths.
