+++ pq_graph/src/fusion.cc
@@ -25,7 +25,7 @@
 #include <iostream>
 
 #include "../include/pq_graph.h"
-#include "../include/code_printer.h"
+#include "../include/printers/code_printer.h"
 
 // include omp only if defined
 #ifdef _OPENMP
