+++ pq_graph/src/graph_printing.cc
@@ -30,8 +30,7 @@
 
 #include "../include/pq_graph.h"
 #include "../include/term.h"
-#include "../include/code_printer.h"
-#include "../include/einsum_printer.h"
+#include "../include/printers/code_printer.h"
 
 
 // include omp only if defined
