+++ pq_graph/src/vertex_printing.cc
@@ -31,7 +31,7 @@
 
 #include "../include/pq_graph.h"
 #include "../include/term.h"
-#include "../include/code_printer.h"
+#include "../include/printers/code_printer.h"
 
 
 // include omp only if defined
